refactored arguments. restructed argument folder to contain cli folder. created a common input parser class
