+++ src/service/arguments/cli/tools/tool-subcommand-factory.ts
@@ -1,8 +1,10 @@
 import { Command } from "commander";
-import { CommandConstructor } from "@bc/service/arguments/command-constructor";
+import { CommandConstructor } from "@bc/service/arguments/cli/command-constructor";
 import { CLIActionType, ToolType } from "@bc/domain/cli";
-import { ProjectListCommand } from "@bc/service/arguments/tools/project-list";
-import { ParsedOptions } from "@bc/service/arguments/parsed-options";
+import { ProjectListCommand } from "@bc/service/arguments/cli/tools/project-list";
+import { ParsedInputs } from "@bc/service/inputs/parsed-inputs";
+import Container from "typedi";
+
 
 
 /**
@@ -29,8 +31,8 @@ export class ToolSubCommandFactory {
             .option("-t, --token <token>", "The GITHUB_TOKEN. It can be set as an environment variable instead")
             .option("-d, --debug", "Set debugging mode to true", false)
             .action((options) => {
-                ParsedOptions.setOpts(options);
-                ParsedOptions.setExecutedCommand({command: CLIActionType.TOOLS, action: toolType});
+                const parsedInputs = Container.get(ParsedInputs);
+                parsedInputs.updateInputs({...options, CLICommand: CLIActionType.TOOLS, CLISubCommand: toolType});
             });
         
     }
